Move exercise detail component to correct folder
