+++ src/components/Exercises/Detail/ExerciseDetails.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { act, render, screen } from '@testing-library/react';
-import { ExerciseDetails } from './index';
-import { MemoryRouter, Route, Routes } from 'react-router';
-import { getExerciseBase, getExerciseBasesForVariation, getLanguageByShortName, getLanguages } from "services";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import {act, render, screen} from '@testing-library/react';
+import {ExerciseDetails} from './ExerciseDetails';
+import {MemoryRouter, Route, Routes} from 'react-router';
+import {getExerciseBase, getExerciseBasesForVariation, getLanguageByShortName, getLanguages} from "services";
+import {QueryClient, QueryClientProvider} from "@tanstack/react-query";
 import {
     testExerciseCrunches,
     testExerciseCurls,
@@ -11,7 +11,7 @@ import {
     testLanguageEnglish,
     testLanguages
 } from "tests/exerciseTestdata";
-import { useLanguageQuery } from "components/Exercises/queries";
+import {useLanguageQuery} from "components/Exercises/queries";
 
 jest.mock("services");
 jest.mock("components/Exercises/queries");
@@ -50,7 +50,7 @@ describe("Render tests", () => {
             <QueryClientProvider client={queryClient}>
                 <MemoryRouter initialEntries={['/exercises/9']}>
                     <Routes>
-                        <Route path='exercises/:baseID' element={<ExerciseDetails />} />
+                        <Route path='exercises/:baseID' element={<ExerciseDetails/>}/>
                     </Routes>
                 </MemoryRouter>
             </QueryClientProvider>
