+++ src/components/Exercises/Detail/Head/head.module.scss
