+++ src/components/Exercises/Detail/Head/index.tsx
