+++ src/components/Exercises/Detail/SideGallery/index.tsx
