+++ src/components/Exercises/Detail/SideGallery/side_gallery.module.scss
