+++ src/components/Exercises/Detail/exerciseDetails.module.css
