+++ src/components/Exercises/Detail/exerciseDetails.module.scss
