+++ src/pages/AddExercise/index.tsx
@@ -1,10 +1,8 @@
 import React from 'react';
-import { AddExerciseStepper } from "components/Exercises/Add/AddExerciseStepper";
+import {AddExerciseStepper} from "components/Exercises/Add/AddExerciseStepper";
 
 export const AddExercise = () => {
     return (
-        <>
-            <AddExerciseStepper />
-        </>
+        <AddExerciseStepper/>
     );
 };
