+++ src/pages/index.ts
@@ -1,19 +1,19 @@
-export { About } from './About';
-export { AddExercise } from './AddExercise';
-export { AddWeight } from './AddWeight';
-export { BmiCalculator } from './BmiCalculator';
-export { Calendar } from './Calendar';
-export { CaloriesCalculator } from './CaloriesCalculator';
-export { Equipments } from './Equipments';
-export { Gallery } from './Gallery';
-export { Ingredients } from './Ingrdedients';
-export { Login } from './Login';
-export { ExerciseDetails } from './ExerciseDetails';
-export { NutritionPlans } from './NutritionPlans';
-export { Preferences } from './Preferences';
-export { PublicTemplate } from './PublicTemplate';
-export { ApiPage } from './ApiPage';
-export { TemplatePage } from './TemplatePage';
-export { WeightOverview } from './WeightOverview';
-export { Workout } from './Workout';
-export { WorkoutSchedule } from './WorkoutSchedule';
+export {About} from './About';
+export {AddExercise} from './AddExercise';
+export {AddWeight} from './AddWeight';
+export {BmiCalculator} from './BmiCalculator';
+export {Calendar} from './Calendar';
+export {CaloriesCalculator} from './CaloriesCalculator';
+export {Equipments} from './Equipments';
+export {Gallery} from './Gallery';
+export {Ingredients} from './Ingrdedients';
+export {Login} from './Login';
+export {ExerciseDetails} from '../components/Exercises/Detail/ExerciseDetails';
+export {NutritionPlans} from './NutritionPlans';
+export {Preferences} from './Preferences';
+export {PublicTemplate} from './PublicTemplate';
+export {ApiPage} from './ApiPage';
+export {TemplatePage} from './TemplatePage';
+export {WeightOverview} from './WeightOverview';
+export {Workout} from './Workout';
+export {WorkoutSchedule} from './WorkoutSchedule';
